refactor: 토큰 기능 추상화
