+++ src/main/java/roomescape/application/service/TokenService.java
@@ -1,0 +1,10 @@
+package roomescape.application.service;
+
+import roomescape.domain.model.Member;
+
+public interface TokenService {
+
+    String createToken(final Member member);
+
+    String checkByToken(String token);
+}
