+++ src/main/resources/application.properties
@@ -1,3 +1,4 @@
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
 spring.datasource.url=jdbc:h2:mem:database
+jwt.secret.key=Yn2kjibddFAWtnPJ2AFlL8WXmohJMCvigQggaEypa5E=
